chore: déplacement de l'interface IUtilisateur dans le dossier Repositories>Interfaces
